:sparkles: feat(frontend): #300 全项收尾——Range 值语义 + A/B/D 修复 + F 解析修复
I 项：Range 是一等值（三槽 [start,end,step]，复用 Tuple 载体）
- r = 1..10 合法；x in r / for i in r 变量形态按静态类型脱糖
- step 形态 a..b..c（双点无新关键词）；step=0 字面量编译期 E1002，
  动态运行时 E6001（ponytail 锚点：未来升格 Result，#301）
- 负 step 递减合法（界方向随符号反转）
- ir_gen 补 BinOp::Range 构造臂，消灭 E3005 unhandled binary operator

A 项：in 类型检查补元素-容器兼容性校验（"a" in [1,2,3] 编译期 E1002）
B 项：Contains 运行时 _ => false 改内部错误；String 分支非 String/Char 探针报错
D 项：IndexOutOfBounds.index usize→i64，l[-1] 报 valid range is 0..3, found -1
F 项：BP_RANGE (6,7)——0..n+2 解析为 0..(n+2)，a..b..c 为 step 形态
G 项：type_name catch-all 已删（046996e8），勾销零改动

文档：syntax.md §2.2 对照表、§2.12 RangeExpr 文法、stdlib.md §6.2 删 step 关键词
E2E：range_value / range_step_zero_err / in_elem_type_err / in_non_container_err / array_neg_index_err

挂账：#301（错误系统，step=0 Result 化）、#302（std.Range + Iterator 落地）

+++ src/backends/mod.rs
@@ -64,8 +64,8 @@ pub enum ExecutorError {
     IndexOutOfBounds {
         /// 容器长度
         max: usize,
-        /// 越界索引
-        index: usize,
+        /// 越界索引（#300 D 项：i64 保真——负索引保留原始值，不再是 max=0 哨兵）
+        index: i64,
         /// 调用栈
         stack: Option<Vec<StackFrame>>,
     },
@@ -166,7 +166,7 @@ impl ExecutorError {
     /// Create an index out of bounds error
     pub fn index_out_of_bounds(
         max: usize,
-        index: usize,
+        index: i64,
         stack: Option<Vec<StackFrame>>,
     ) -> Self {
         ExecutorError::IndexOutOfBounds { max, index, stack }
